+++ internal/controller/homeassistantsecrets_controller.go
@@ -64,7 +64,7 @@ type HomeAssistantSecretsReconciler struct {
 // +kubebuilder:rbac:groups=ha.homeassistant.io,resources=homeassistantsecrets,verbs=get;list;watch
 // +kubebuilder:rbac:groups=ha.homeassistant.io,resources=homeassistantsecrets/status,verbs=get;update;patch
 // +kubebuilder:rbac:groups=ha.homeassistant.io,resources=homeassistantsecrets/finalizers,verbs=update
-// +kubebuilder:rbac:groups=core,resources=secrets,verbs=get;list;watch;create;update;patch;delete
+// +kubebuilder:rbac:groups=core,resources=secrets,verbs=get;list;watch;create;update
 // +kubebuilder:rbac:groups=ha.homeassistant.io,resources=homeassistants,verbs=get;list;watch
 // +kubebuilder:rbac:groups=apps,resources=statefulsets,verbs=get;list;watch;update;patch